src/requirements.txt
pytubefix
python-ffmpeg==2.0.12
StrEnum==0.4.15; python_version < '3.11'